Adding generic driver code (WIP)
This is part of a larger set of patches, This contains the "RouterDriver".

This patch also includes a naming convention change:
router_id = instance_id
vm = instance
VMManager = instance_manager

+++ akanda/rug/instance_manager.py
@@ -1,4 +1,4 @@
-# Copyright 2014 DreamHost, LLC
+# Copyright 2015 DreamHost, LLC
 #
 # Author: DreamHost, LLC
 #
@@ -14,14 +14,12 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
-
 from datetime import datetime
 from functools import wraps
 import time
 
 from oslo.config import cfg
 
-from akanda.rug.api import configuration
 from akanda.rug.api import akanda_client as router_api
 from akanda.rug.api import neutron
 
@@ -43,17 +41,19 @@ STATUS_MAP = {
 }
 
 
-def synchronize_router_status(f):
+def synchronize_instance_status(f):
+    """decorator for syncing instance status"""
     @wraps(f)
     def wrapper(self, worker_context, silent=False):
         old_status = self._last_synced_status
         val = f(self, worker_context, silent)
-        if not self.router_obj:
+        if not self.instance_obj:
             return val
         new_status = STATUS_MAP.get(self.state, neutron.STATUS_ERROR)
         if not old_status or old_status != new_status:
-            worker_context.neutron.update_router_status(
-                self.router_obj.id,
+            self.driver.update_status(
+                worker_context,
+                self.instance_obj,
                 new_status
             )
             self._last_synced_status = new_status
@@ -62,6 +62,7 @@ def synchronize_router_status(f):
 
 
 class BootAttemptCounter(object):
+    """Boot attempt counter object"""
     def __init__(self):
         self._attempts = 0
 
@@ -76,14 +77,14 @@ class BootAttemptCounter(object):
         return self._attempts
 
 
-class VmManager(object):
-
-    def __init__(self, router_id, tenant_id, log, worker_context):
-        self.router_id = router_id
+class InstanceManager(object):
+    def __init__(self, instance_id, tenant_id, log, driver, worker_context):
+        self.instance_id = instance_id
         self.tenant_id = tenant_id
         self.log = log
+        self.driver = driver
         self.state = DOWN
-        self.router_obj = None
+        self.instance_obj = None
         self.instance_info = None
         self.last_error = None
         self._boot_counter = BootAttemptCounter()
@@ -97,15 +98,15 @@ class VmManager(object):
     def reset_boot_counter(self):
         self._boot_counter.reset()
 
-    @synchronize_router_status
+    @synchronize_instance_status
     def update_state(self, worker_context, silent=False):
         self._ensure_cache(worker_context)
         if self.state == GONE:
-            self.log.debug('not updating state of deleted router')
+            self.log.debug('not updating state of deleted instance')
             return self.state
 
         if self.instance_info is None:
-            self.log.debug('no backing instance, marking router as down')
+            self.log.debug('no backing instance, marking instance as down')
             self.state = DOWN
             return self.state
 
@@ -126,23 +127,23 @@ class VmManager(object):
             old_state = self.state
             self._check_boot_timeout()
 
-            # If the router isn't responding, make sure Nova knows about it
+            # If the instance isn't responding, make sure Nova knows about it
             instance = worker_context.nova_client.get_instance_for_obj(
-                self.router_id
+                self.instance_id
             )
             if instance is None and self.state != ERROR:
-                self.log.info('No router VM was found; rebooting')
+                self.log.info('No VM was found; rebooting')
                 self.state = DOWN
                 self.instance_info = None
 
             # update_state() is called from Alive() to check the
-            # status of the router. If we can't talk to the API at
-            # that point, the router should be considered missing and
+            # status of the instance. If we can't talk to the API at
+            # that point, the instance should be considered missing and
             # we should reboot it, so mark it down if we think it was
             # configured before.
             if old_state == CONFIGURED and self.state != ERROR:
                 self.log.debug(
-                    'Did not find router alive, marking it as down',
+                    'Did not find instance alive, marking it as down',
                 )
                 self.state = DOWN
 
@@ -154,7 +155,8 @@ class VmManager(object):
             # duration to log.
             self.instance_info.confirm_up()
             if self.instance_info.boot_duration:
-                self.log.info('Router booted in %s seconds after %s attempts',
+                self.log.info('%s booted in %s seconds after %s attempts',
+                              self.driver.RESOURCE_NAME,
                               self.instance_info.boot_duration.total_seconds(),
                               self._boot_counter.count)
             # Always reset the boot counter, even if we didn't boot
@@ -163,46 +165,52 @@ class VmManager(object):
             self._boot_counter.reset()
         return self.state
 
-    def boot(self, worker_context, router_image_uuid):
+    def boot(self, worker_context):
         self._ensure_cache(worker_context)
         if self.state == GONE:
-            self.log.info('not booting deleted router')
+            self.log.info('not booting deleted %s', self.driver.RESOURCE_NAME)
             return
 
-        self.log.info('Booting router')
+        self.log.info('Booting %s', self.driver.RESOURCE_NAME)
         self.state = DOWN
         self._boot_counter.start()
 
         def make_vrrp_ports():
             mgt_port = worker_context.neutron.create_management_port(
-                self.router_obj.id
+                self.instance_obj.id
             )
 
             # FIXME(mark): ideally this should be ordered and de-duped
             instance_ports = [
-                worker_context.neutron.create_vrrp_port(self.router_obj.id, n)
-                for n in (p.network_id for p in self.router_obj.ports)
+                worker_context.neutron.create_vrrp_port(self.instance_obj.id,
+                                                        n)
+                for n in (p.network_id for p in self.instance_obj.ports)
             ]
 
             return mgt_port, instance_ports
 
         try:
-            # TODO(mark): make this pluggable
-            self._ensure_provider_ports(self.router_obj, worker_context)
+            self.driver.preboot(worker_context, self.instance_obj)
 
             # TODO(mark): make this handle errors more gracefully on cb fail
             # TODO(mark): checkout from a pool - boot on demand for now
             instance_info = worker_context.nova_client.boot_instance(
                 self.instance_info,
-                self.router_obj.id,
-                router_image_uuid,
+                self.instance_obj.id,
+                self.driver.image_uuid,
                 make_vrrp_ports
             )
             if not instance_info:
-                self.log.info('Previous router is deleting')
+                self.log.info(
+                    'Previous %s is deleting',
+                    self.driver.RESOURCE_NAME
+                )
                 return
         except:
-            self.log.exception('Router failed to start boot')
+            self.log.exception(
+                '%s failed to start boot',
+                self.driver.RESOURCE_NAME
+            )
             # TODO(mark): attempt clean-up of failed ports
             return
         else:
@@ -214,16 +222,23 @@ class VmManager(object):
     def check_boot(self, worker_context):
         ready_states = (UP, CONFIGURED)
         if self.update_state(worker_context, silent=True) in ready_states:
-            self.log.info('Router has booted, attempting initial config')
+            self.log.info(
+                '%s has booted, attempting initial config',
+                self.driver.RESOURCE_NAME
+            )
             self.configure(worker_context, BOOTING, attempts=1)
             if self.state != CONFIGURED:
                 self._check_boot_timeout()
             return self.state == CONFIGURED
 
-        self.log.debug('Router is %s' % self.state.upper())
+        self.log.debug(
+            '%s is %s',
+            self.driver.RESOURCE_NAME,
+            self.state.upper()
+        )
         return False
 
-    @synchronize_router_status
+    @synchronize_instance_status
     def set_error(self, worker_context, silent=False):
         """Set the internal and neutron status for the router to ERROR.
 
@@ -234,13 +249,16 @@ class VmManager(object):
         """
         self._ensure_cache(worker_context)
         if self.state == GONE:
-            self.log.debug('not updating state of deleted router')
+            self.log.debug(
+                'not updating state of deleted %s',
+                self.driver.RESOURCE_NAME
+            )
             return self.state
         self.state = ERROR
         self.last_error = datetime.utcnow()
         return self.state
 
-    @synchronize_router_status
+    @synchronize_instance_status
     def clear_error(self, worker_context, silent=False):
         """Clear the internal error state.
 
@@ -253,7 +271,10 @@ class VmManager(object):
         self._boot_counter.reset()
         self._ensure_cache(worker_context)
         if self.state == GONE:
-            self.log.debug('not updating state of deleted router')
+            self.log.debug(
+                'not updating state of deleted %s',
+                self.driver.RESOURCE_NAME
+            )
             return self.state
         self.state = DOWN
         return self.state
@@ -272,15 +293,28 @@ class VmManager(object):
     def stop(self, worker_context):
         self._ensure_cache(worker_context)
         if self.state == GONE:
-            self.log.info('Destroying router neutron has deleted')
+            # We are being told to delete an instance that neutron has
+            # already removed.
+            self.log.info(
+                'Destroying %s neutron has deleted',
+                self.driver.RESOURCE_NAME
+            )
         else:
-            self.log.info('Destroying router')
+            self.log.info('Destroying %s', self.driver.RESOURCE_NAME)
+
+        # TODO(mark): remove this dev hook or make it real
+        if hasattr(self.driver, 'DONTSTOP'):
+            self.state = ERROR
+            return
 
         try:
             nova_client = worker_context.nova_client
             nova_client.destroy_instance(self.instance_info)
         except Exception:
-            self.log.exception('Error deleting router instance')
+            self.log.exception(
+                'Error deleting %s instance',
+                self.driver.RESOURCE_NAME
+            )
 
         start = time.time()
         while time.time() - start < cfg.CONF.boot_timeout:
@@ -288,14 +322,18 @@ class VmManager(object):
                 if self.state != GONE:
                     self.state = DOWN
                 return
-            self.log.debug('Router has not finished stopping')
+            self.log.debug(
+                '%s has not finished stopping',
+                self.driver.RESOURCE_NAME
+            )
             time.sleep(cfg.CONF.retry_delay)
         self.log.error(
-            'Router failed to stop within %d secs',
+            '%sfailed to stop within %d secs',
+            self.driver.RESOURCE_NAME,
             cfg.CONF.boot_timeout)
 
     def configure(self, worker_context, failure_state=RESTART, attempts=None):
-        self.log.debug('Begin router config')
+        self.log.debug('Begin %s config', self.driver.RESOURCE_NAME)
         self.state = UP
         attempts = attempts or cfg.CONF.max_retries
 
@@ -316,7 +354,7 @@ class VmManager(object):
             cfg.CONF.akanda_mgt_service_port
         )
 
-        if not self._verify_interfaces(self.router_obj, interfaces):
+        if not self._verify_interfaces(self.instance_obj, interfaces):
             # FIXME: Need a REPLUG state when we support hot-plugging
             # interfaces.
             self.log.debug("Interfaces aren't plugged as expected.")
@@ -341,12 +379,13 @@ class VmManager(object):
         }
 
         # FIXME: Need to catch errors talking to neutron here.
-        config = configuration.build_config(
-            worker_context.neutron,
-            self.router_obj,
-            mgt_port,
+        config = self.driver.build_config(
+            worker_context,
+            self.instance_obj,
+            self.instance_info,
             iface_map
         )
+
         self.log.debug('preparing to update config to %r', config)
 
         for i in xrange(attempts):
@@ -377,7 +416,7 @@ class VmManager(object):
 
     def replug(self, worker_context):
         self.log.debug('Attempting to replug...')
-        self._ensure_provider_ports(self.router_obj, worker_context)
+        self.driver.preplug(worker_context, self.instance_obj)
 
         interfaces = router_api.get_interfaces(
             self.instance_info.management_address,
@@ -400,7 +439,7 @@ class VmManager(object):
         instance_ports = {p.network_id: p for p in self.instance_info.ports}
         instance_networks = set(instance_ports.keys())
 
-        logical_networks = set(p.network_id for p in self.router_obj.ports)
+        logical_networks = set(p.network_id for p in self.instance_obj.ports)
 
         if logical_networks != instance_networks:
             instance = worker_context.nova_client.get_instance_by_id(
@@ -410,7 +449,7 @@ class VmManager(object):
             # For each port that doesn't have a mac address on the VM...
             for network_id in logical_networks - instance_networks:
                 port = worker_context.neutron.create_vrrp_port(
-                    self.router_obj.id,
+                    self.instance_obj.id,
                     network_id
                 )
                 self.log.debug(
@@ -445,7 +484,7 @@ class VmManager(object):
                 self.instance_info.management_address,
                 cfg.CONF.akanda_mgt_service_port
             )
-            if self._verify_interfaces(self.router_obj, interfaces):
+            if self._verify_interfaces(self.instance_obj, interfaces):
                 # replugging was successful
                 # TODO(mark) update port states
                 return
@@ -457,19 +496,20 @@ class VmManager(object):
 
     def _ensure_cache(self, worker_context):
         try:
-            self.router_obj = worker_context.neutron.get_router_detail(
-                self.router_id
+            self.instance_obj = self.driver.get_logical_config(
+                worker_context,
+                self.instance_id
             )
-        except neutron.RouterGone:
+        except neutron.LogicalInstanceGone:
             # The router has been deleted, set our state accordingly
             # and return without doing any more work.
             self.state = GONE
-            self.router_obj = None
+            self.instance_obj = None
 
         if not self.instance_info:
             self.instance_info = (
                 worker_context.nova_client.get_instance_info_for_obj(
-                    self.router_id
+                    self.instance_id
                 )
             )
 
@@ -525,14 +565,3 @@ class VmManager(object):
         self.log.debug('MACs expected: %s', ', '.join(sorted(expected_macs)))
 
         return router_macs == expected_macs
-
-    def _ensure_provider_ports(self, router, worker_context):
-        if router.external_port is None:
-            # FIXME: Need to do some work to pick the right external
-            # network for a tenant.
-            self.log.debug('Adding external port to router')
-            ext_port = worker_context.neutron.create_router_external_port(
-                router
-            )
-            router.external_port = ext_port
-        return router
